003 Code/dyslexia-ai/requirements.txt
# Web Framework & API
fastapi[all]~=0.115.14
python-multipart~=0.0.20

# AI & LLM
langchain~=0.3.26
langchain-anthropic~=0.3.16
anthropic~=0.56.0
langchain-openai~=0.2.11
openai~=1.57.0
replicate~=1.0.7

# Document Processing
pdfplumber~=0.11.7
jpype1~=1.5.2
konlpy~=0.6.0

# Environment Variables
python-dotenv~=1.1.1

# HTTP Requests
requests~=2.31.0

# Redis Cache (비동기 지원 포함)
redis~=6.1.1

# AWS Services
boto3~=1.39.4

# HTTP Client for Webhooks
httpx~=0.27.0
